+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/games/GameEmulator.java
@@ -2,29 +2,18 @@ package de.mephisto.vpin.server.games;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import de.mephisto.vpin.restclient.popper.Emulator;
-import de.mephisto.vpin.server.roms.RomService;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
-import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.nio.charset.Charset;
-import java.util.*;
-import java.util.stream.Collectors;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 public class GameEmulator {
-  private final static Logger LOG = LoggerFactory.getLogger(RomService.class);
-
-  private Map<String, String> aliasToRomMapping = new HashMap<>();
-
   private final static String VPREG_STG = "VPReg.stg";
-  private final static String VPM_ALIAS = "VPMAlias.txt";
 
   private final File installationFolder;
   private final File tablesFolder;
@@ -85,8 +74,6 @@ public class GameEmulator {
     if (!StringUtils.isEmpty(emulator.getDirRoms())) {
       this.romFolder = new File(emulator.getDirRoms());
     }
-
-    loadAliasMapping();
   }
 
   public boolean isVisible() {
@@ -123,23 +110,6 @@ public class GameEmulator {
 
   public String getTablesDirectory() {
     return tablesDirectory;
-  }
-
-  public String getRomForAlias(@Nullable String romAlias) {
-    if (romAlias == null) {
-      return null;
-    }
-
-    Set<Map.Entry<String, String>> entries = aliasToRomMapping.entrySet();
-    for (Map.Entry<String, String> entry : entries) {
-      String alias = entry.getKey();
-      String romName = entry.getValue();
-
-      if (alias.equals(romAlias)) {
-        return romName;
-      }
-    }
-    return null;
   }
 
   public String getName() {
@@ -240,84 +210,6 @@ public class GameEmulator {
   @JsonIgnore
   public File getTablesFolder() {
     return tablesFolder;
-  }
-
-  @NonNull
-  @JsonIgnore
-  public File getVPMAliasFile() {
-    return new File(mameFolder, VPM_ALIAS);
-  }
-
-
-  public boolean clearCache() {
-    this.aliasToRomMapping.clear();
-    this.loadAliasMapping();
-    return true;
-  }
-
-  public boolean deleteAliasMapping(String alias) throws IOException {
-    if (!StringUtils.isEmpty(alias) && aliasToRomMapping.containsKey(alias)) {
-      aliasToRomMapping.remove(alias);
-      LOG.info("Removed alias mapping '" + alias + "'");
-      saveMapping();
-    }
-    return false;
-  }
-
-  public boolean saveAliasMapping(Map<String, Object> values) throws IOException {
-    String oldValue = (String) values.get("#oldValue");
-    if (!StringUtils.isEmpty(oldValue) && aliasToRomMapping.containsKey(oldValue)) {
-      aliasToRomMapping.remove(oldValue);
-      LOG.info("Removed old alias mapping '" + oldValue + "'");
-    }
-    values.remove("#oldValue");
-
-
-    Set<Map.Entry<String, Object>> entries = values.entrySet();
-    for (Map.Entry<String, Object> entry : entries) {
-      String alias = entry.getKey();
-      String romName = (String) entry.getValue();
-
-      if (!aliasToRomMapping.containsValue(alias)) {
-        aliasToRomMapping.put(alias.trim(), romName.trim());
-      }
-    }
-
-    saveMapping();
-    return true;
-  }
-
-  private void saveMapping() throws IOException {
-    String mapAsString = aliasToRomMapping.keySet().stream().map(key -> key.trim() + "," + aliasToRomMapping.get(key).trim()).sorted().collect(Collectors.joining("\n"));
-    File vpmAliasFile = getVPMAliasFile();
-    FileUtils.writeStringToFile(vpmAliasFile, mapAsString, Charset.defaultCharset());
-    loadAliasMapping();
-  }
-
-  private void loadAliasMapping() {
-    File vpmAliasFile = getVPMAliasFile();
-    try {
-      if (vpmAliasFile.exists()) {
-        FileInputStream fileInputStream = new FileInputStream(vpmAliasFile);
-        List<String> mappings = IOUtils.readLines(fileInputStream, "utf8");
-        fileInputStream.close();
-
-        for (String mapping : mappings) {
-          if (mapping.contains(",")) {
-            String[] split = mapping.split(",");
-            String[] aliases = Arrays.copyOfRange(split, 0, split.length - 1);
-            String originalRom = split[split.length - 1];
-
-            for (String alias : aliases) {
-              aliasToRomMapping.put(alias, originalRom);
-            }
-          }
-        }
-        LOG.info("Loaded " + aliasToRomMapping.size() + " alias mappings.");
-      }
-    } catch (IOException e) {
-      LOG.error("Error loading " + vpmAliasFile.getAbsolutePath() + ": " + e.getMessage(), e);
-    }
   }
 
   @Override